+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/telemetry
 
+## 2.0.6
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.13 → 2.2.0
+
+## 2.0.5
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/nodejs-utils 1.0.9 → 1.0.10
+
 ## 2.0.4
 
 ### Patch Changes