+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 1.0.41
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.13 → 2.2.0
+- @sap-ux/fiori-generator-shared 1.2.14 → 1.2.15
+- @sap-ux/telemetry 2.0.5 → 2.0.6
+- @sap-ux/ui5-library-reference-inquirer 1.0.40 → 1.0.41
+- @sap-ux/ui5-library-reference-writer 1.0.18 → 1.0.19
+
+## 1.0.40
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.2.13 → 1.2.14
+- @sap-ux/ui5-library-reference-inquirer 1.0.39 → 1.0.40
+
+## 1.0.39
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/telemetry 2.0.4 → 2.0.5
+- @sap-ux/fiori-generator-shared 1.2.12 → 1.2.13
+- @sap-ux/ui5-library-reference-inquirer 1.0.38 → 1.0.39
+
 ## 1.0.38
 
 ### Patch Changes